linera_parity_wasm::elements

Struct Module

pub struct Module { /* private fields */ }

WebAssembly module

Implementations

impl Module

pub fn new(sections: Vec<Section>) -> Self

New module with sections

pub fn from_bytes<T: AsRef<[u8]>>(input: T) -> Result<Self, Error>

Construct a module from a slice.

pub fn into_bytes(self) -> Result<Vec<u8>, Error>

Serialize a module to a vector.

pub fn into_sections(self) -> Vec<Section>

Destructure the module, yielding sections

pub fn version(&self) -> u32

Version of module.

pub fn sections(&self) -> &[Section]

Sections list.

Each known section is optional and may appear at most once.

pub fn sections_mut(&mut self) -> &mut Vec<Section>

Sections list (mutable).

Each known section is optional and may appear at most once.

pub fn insert_section(&mut self, section: Section) -> Result<(), Error>

Insert a section, in the correct section ordering. This will fail with an error, if the section can only appear once.

pub fn code_section(&self) -> Option<&CodeSection>

Code section reference, if any.

pub fn code_section_mut(&mut self) -> Option<&mut CodeSection>

Code section mutable reference, if any.

pub fn type_section(&self) -> Option<&TypeSection>

Types section reference, if any.

pub fn type_section_mut(&mut self) -> Option<&mut TypeSection>

Types section mutable reference, if any.

pub fn import_section(&self) -> Option<&ImportSection>

Imports section reference, if any.

pub fn import_section_mut(&mut self) -> Option<&mut ImportSection>

Imports section mutable reference, if any.

pub fn global_section(&self) -> Option<&GlobalSection>

Globals section reference, if any.

pub fn global_section_mut(&mut self) -> Option<&mut GlobalSection>

Globals section mutable reference, if any.

pub fn export_section(&self) -> Option<&ExportSection>

Exports section reference, if any.

pub fn export_section_mut(&mut self) -> Option<&mut ExportSection>

Exports section mutable reference, if any.

pub fn table_section(&self) -> Option<&TableSection>

Table section reference, if any.

pub fn table_section_mut(&mut self) -> Option<&mut TableSection>

Table section mutable reference, if any.

pub fn data_section(&self) -> Option<&DataSection>

Data section reference, if any.

pub fn data_section_mut(&mut self) -> Option<&mut DataSection>

Data section mutable reference, if any.

pub fn elements_section(&self) -> Option<&ElementSection>

Element section reference, if any.

pub fn elements_section_mut(&mut self) -> Option<&mut ElementSection>

Element section mutable reference, if any.

pub fn memory_section(&self) -> Option<&MemorySection>

Memory section reference, if any.

pub fn memory_section_mut(&mut self) -> Option<&mut MemorySection>

Memory section mutable reference, if any.

pub fn function_section(&self) -> Option<&FunctionSection>

Functions signatures section reference, if any.

pub fn function_section_mut(&mut self) -> Option<&mut FunctionSection>

Functions signatures section mutable reference, if any.

pub fn start_section(&self) -> Option<u32>

Start section, if any.

pub fn set_start_section(&mut self, new_start: u32)

Changes the module’s start section.

pub fn clear_start_section(&mut self)

Removes the module’s start section.

pub fn custom_sections(&self) -> impl Iterator<Item = &CustomSection>

Returns an iterator over the module’s custom sections

pub fn set_custom_section(&mut self, name: impl Into<String>, payload: Vec<u8>)

Sets the payload associated with the given custom section, or adds a new custom section, as appropriate.

pub fn clear_custom_section( &mut self, name: impl AsRef<str>, ) -> Option<CustomSection>

Removes the given custom section, if it exists. Returns the removed section if it existed, or None otherwise.

pub fn has_names_section(&self) -> bool

True if a name section is present.

NOTE: this can return true even if the section was not parsed, hence names_section() may return None even if this returns true

pub fn names_section(&self) -> Option<&NameSection>

Functions signatures section reference, if any.

NOTE: name section is not parsed by default so names_section could return None even if name section exists. Call parse_names to parse name section

pub fn names_section_mut(&mut self) -> Option<&mut NameSection>

Functions signatures section mutable reference, if any.

NOTE: name section is not parsed by default so names_section could return None even if name section exists. Call parse_names to parse name section

pub fn parse_names(self) -> Result<Self, (Vec<(usize, Error)>, Self)>

Try to parse name section in place.

Corresponding custom section with proper header will convert to name sections If some of them will fail to be decoded, Err variant is returned with the list of (index, Error) tuples of failed sections.

pub fn parse_reloc(self) -> Result<Self, (Vec<(usize, Error)>, Self)>

Try to parse reloc section in place.

Corresponding custom section with proper header will convert to reloc sections If some of them will fail to be decoded, Err variant is returned with the list of (index, Error) tuples of failed sections.

pub fn import_count(&self, count_type: ImportCountType) -> usize

Count imports by provided type.

pub fn functions_space(&self) -> usize

Query functions space.

pub fn globals_space(&self) -> usize

Query globals space.

pub fn table_space(&self) -> usize

Query table space.

pub fn memory_space(&self) -> usize

Query memory space.